Add files via upload
This is a Database project implemented using ORACLE Database. In this project, I have developed a database for a hotel management system.

Project Details:
The database contains a large dataset of 190,000 entries, with 10,000 entries in each table, as per project requirements.
There are 19 tables in the database.
All necessary files and queries are included in the provided RAR file.

To Run the Project:
Copy and paste the queries into your Oracle Database.

Note that the queries are specifically written for Oracle Database.

I have used the SQL Developer IDE/Compiler to run the project.